feat: Add build and run configurations to `pom.xml`
This commit adds Maven configurations to build a runnable JAR file
and to directly execute the application using Maven.

The `maven-jar-plugin` is configured to create a JAR with dependencies
included and specifies the main class to execute when running the JAR.

The `exec-maven-plugin` allows running the application directly through
Maven, simplifying development and testing.